requirements.txt
einops==0.8.0
h5py==3.11.0
matplotlib==3.7.0
numpy==2.1.3
omegaconf==2.3.0
opencv_python==4.7.0.72
pandas==1.5.3
Pillow==11.0.0
scikit_learn==1.3.2
scipy==1.14.1
seaborn
setuptools==60.2.0
scikit-image
tensorboardX==2.6.2.2
tensorboardX==2.6.2.2
torch==2.3.0+cu121
torchvision==0.18.0
tqdm==4.65.2
